feat: Add image dimensions to photo form
This commit adds the display of image dimensions to the photo form. It includes the necessary changes to the index.html file to add two new paragraphs with ids "imageWidth" and "imageHeight". The dimensions are dynamically updated with the width and height of the captured image. This enhancement provides users with information about the size of the photo they are about to submit.